requirements.txt
asgiref==3.2.9
autopep8==1.5.3
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
codecov==2.1.7
coverage==5.1
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.7
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
Markdown==3.2.2
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
whitenoise==5.1.0
